+++ db/migrate/20210901084543_create_claim_decisions.rb
@@ -1,0 +1,5 @@
+class CreateClaimDecisions < ActiveRecord::Migration[6.0]
+  def change
+    create_view :claim_decisions
+  end
+end