+++ CHANGELOG.md
@@ -99,6 +99,7 @@ The format is based on [Keep a Changelog]
   - Non-descriptive links
 - Google Analytics
   - Add Google Tag Manager to admin and application
+  - Adds CSV view for claims decsions
 - Claims
   - Stop claim full name adding redundant spaces when middle name is an empty or
     blank string